// server/src/types.ts
export interface PistonExecuteResponse {
  language: string;
  version: string;
  run: {
    stdout: string;
    stderr: string;
    code: number;
    signal: string | null;
    output: string;
    memory?: number;
    cpu_time?: number;
    wall_time?: number;
  };
  compile?: {
    stdout: string;
    stderr: string;
    code: number;
    signal: string | null;
    output: string;
    memory?: number;
    cpu_time?: number;
    wall_time?: number;
  };
}